[Fix] 상세 정보의 세부 css 조정 및 컴포넌트 이름 변경, import 수정
